For checking!
Missing: ErrorBindingHandler not connected!

+++ src/main/java/com/codecool/workfinder/service/ClientService.java
@@ -3,8 +3,11 @@ package com.codecool.workfinder.service;
 import com.codecool.workfinder.model.dto.ClientDto;
 import com.codecool.workfinder.model.dto.JobDto;
 import com.codecool.workfinder.model.entity.Client;
+import com.codecool.workfinder.model.entity.Job;
 import com.codecool.workfinder.model.mapper.ClientMapper;
+import com.codecool.workfinder.model.mapper.JobMapper;
 import com.codecool.workfinder.repository.ClientRepository;
+import com.codecool.workfinder.repository.JobRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,36 +17,60 @@ import java.util.UUID;
 @Service
 public class ClientService extends BaseService<Client, ClientDto, UUID> {
 
+    private final JobRepository jobRepository;
+    private final JobMapper jobMapper;
     @Autowired
     public ClientService(ClientRepository clientRepository,
-                         ClientMapper clientMapper) {
+                         ClientMapper clientMapper,
+                         JobRepository jobRepository,
+                         JobMapper jobMapper) {
         super(clientRepository, clientMapper);
+        this.jobRepository = jobRepository;
+        this.jobMapper = jobMapper;
     }
 
-    public ClientDto deleteById(UUID uuid) {
+    public Optional<ClientDto> deleteById(UUID uuid) {
+        Client client = repository.findById(uuid).orElse(null);
+        if (client == null) {
+            return Optional.empty();
+        }
         repository.deleteById(uuid);
-        Client client = repository.getById(uuid);
         logInfoViaRepository("Completed accessing repository!");
         ClientDto clientDto = mapper.toDto(client);
         mapper.logInfo("Completed converting to ClientDto!");
         logger.info("Completed method: deleteById(UUID)!");
-        return clientDto;
+        return Optional.of(clientDto);
     }
 
     public void save(ClientDto clientDto) {
-        clientDto.setId(UUID.randomUUID());
+        if (clientDto.getId() == null) {
+            clientDto.setId(UUID.randomUUID());
+        }
         Client client = mapper.toEntity(clientDto);
         mapper.logInfo("Completed converting to Client!");
+        client.getJobs().forEach(job -> {
+            if (job.getId() == null) {
+                job.setId(UUID.randomUUID());
+            }
+
+        });
         repository.save(client);
         logInfoViaRepository("Completed accessing repository!");
         logger.info("Completed method: save(Client)!");
     }
 
-    public Optional<JobDto> registerJobForClient(ClientDto clientDto,
-                                                 JobDto jobDto) {
+    public void registerJobForClient(ClientDto clientDto,
+                                     JobDto jobDto) {
+
         Client client = mapper.toEntity(clientDto);
-        //Job job =
-        return null;
+        mapper.logInfo("Completed converting to Client!");
+        Job job = jobMapper.toEntity(jobDto);
+        client.getJobs().add(job);
+        mapper.logInfo("Completed converting to Job!");
+        jobRepository.save(job);
+        jobRepository.logInfo("Completed accessing repository!");
+        repository.save(client);
+        logInfoViaRepository("Completed accessing repository!");
     }
 
     private void logInfoViaRepository(String message) {
